clase-04/requirements.txt
kafka-python
fastapi
uvicorn[standard]